// src/providers/types/api-private.ts
export type ProfilePerClient = '[U]'|'[C]'|'[M]'|'[T]'|'[MN]'|'[I]'|'[CP]';

export type FaultAction = 'APPROVED'|'REJECTED'|'ERASED'|'RESTAB_WAITING'|'RESTAB_PENDING'|'RESTABLISHED'|'RESTAB_REJECTED'|'PENDING';

export interface UserPermissions {
  MANAGE_ALL_CLIENTS_AND_DACS?: boolean
  PER_CLIENT: { clientId: number, p: ProfilePerClient[] }[]
  PER_UNIT?: { clientId: number, units: number[], p: ProfilePerClient[] }[]
  ALTER_SYSTEM_PARS?: boolean
  VIEW_ALL_CLIENTS_DACS_GROUPS_UNITS?: boolean
  MANAGE_UNOWNED_DEVS?: boolean
  HEALTH_MANAGEMENT?: boolean
  isUserManut?: boolean
  isParceiroValidador?: boolean
  isAdminSistema?: boolean
  isInstaller?: boolean
  API_MANAGEMENT?: boolean
}